Refactor and reorder code
