+++ Center windows winui/Interop/Enums.cs
@@ -1,4 +1,25 @@
-﻿namespace CenterWindow.Interop;
+﻿using System.Runtime.InteropServices;
+
+namespace CenterWindow.Interop;
 internal static partial class NativeMethods
 {
+    public enum PROCESS_ACCESS_TYPES
+    {
+        PROCESS_TERMINATE = 0x00000001,
+        PROCESS_CREATE_THREAD = 0x00000002,
+        PROCESS_SET_SESSIONID = 0x00000004,
+        PROCESS_VM_OPERATION = 0x00000008,
+        PROCESS_VM_READ = 0x00000010,
+        PROCESS_VM_WRITE = 0x00000020,
+        PROCESS_DUP_HANDLE = 0x00000040,
+        PROCESS_CREATE_PROCESS = 0x00000080,
+        PROCESS_SET_QUOTA = 0x00000100,
+        PROCESS_SET_INFORMATION = 0x00000200,
+        PROCESS_QUERY_INFORMATION = 0x00000400,
+        STANDARD_RIGHTS_REQUIRED = 0x000F0000,
+        SYNCHRONIZE = 0x00100000,
+        PROCESS_ALL_ACCESS = PROCESS_TERMINATE | PROCESS_CREATE_THREAD | PROCESS_SET_SESSIONID | PROCESS_VM_OPERATION |
+            PROCESS_VM_READ | PROCESS_VM_WRITE | PROCESS_DUP_HANDLE | PROCESS_CREATE_PROCESS | PROCESS_SET_QUOTA |
+            PROCESS_SET_INFORMATION | PROCESS_QUERY_INFORMATION | STANDARD_RIGHTS_REQUIRED | SYNCHRONIZE
+    }
 }
