+++ Center windows winui/Interop/Gdi32.cs
@@ -1,4 +1,5 @@
 ﻿using System.Runtime.InteropServices;
+using static CenterWindow.Interop.NativeMethods;
 
 namespace CenterWindow.Interop;
 internal static partial class NativeMethod
@@ -10,11 +11,20 @@ internal static partial class NativeMethod
     public static extern IntPtr CreateCompatibleDC(IntPtr hdc);
 
     [DllImport("gdi32.dll", SetLastError = true)]
+    public static extern IntPtr CreateDIBSection(
+        IntPtr hdc,
+        ref BITMAPINFO pbmi,
+        uint usage,
+        out IntPtr ppvBits,
+        IntPtr hSection,
+        uint offset);
+
+    [DllImport("gdi32.dll", SetLastError = true)]
     public static extern bool DeleteDC(IntPtr hdc);
 
     [DllImport("gdi32.dll", SetLastError = true)]
     public static extern bool DeleteObject(IntPtr hObject);
 
     [DllImport("gdi32.dll", SetLastError = true)]
-    public static extern IntPtr SelectObject(IntPtr hdc, IntPtr hgdiobj);   
+    public static extern IntPtr SelectObject(IntPtr hdc, IntPtr hgdiobj);
 }
