+++ Center windows winui/Interop/Structs.cs
@@ -4,60 +4,45 @@ namespace CenterWindow.Interop;
 internal static partial class NativeMethods
 {
     [StructLayout(LayoutKind.Sequential)]
-    public struct Rect
+    public struct BITMAPINFO
     {
-        public int Left;
-        public int Top;
-        public int Right;
-        public int Bottom;
-
-        public readonly int Width => Right - Left;
-        public readonly int Height => Bottom - Top;
+        public BITMAPINFOHEADER bmiHeader;
+        public uint bmiColors;  // espacio mínimo para el color table
     }
 
     [StructLayout(LayoutKind.Sequential)]
-    public struct POINT
+    public struct BITMAPINFOHEADER
     {
-        public int x;
-        public int y;
+        public uint biSize;
+        public int biWidth;
+        public int biHeight;
+        public ushort biPlanes;
+        public ushort biBitCount;
+        public uint biCompression;
+        public uint biSizeImage;
+        public int biXPelsPerMeter;
+        public int biYPelsPerMeter;
+        public uint biClrUsed;
+        public uint biClrImportant;
     }
 
     [StructLayout(LayoutKind.Sequential)]
-    public struct MSLLHOOKSTRUCT
+    public struct GdiplusStartupInput
     {
-        public POINT pt;
-        public uint mouseData;
-        public uint flags;
-        public uint time;
-        public IntPtr dwExtraInfo;
+        public uint GdiplusVersion;
+        public IntPtr DebugEventCallback;
+        public bool SuppressBackgroundThread;
+        public bool SuppressExternalCodecs;
     }
 
-    public struct WindowInfo
+    [StructLayout(LayoutKind.Sequential)]
+    public struct ICONINFO
     {
-        public int size;
-        public Rect window;
-        public Rect client;
-        public int style;
-        public int exStyle;
-        public int windowStatus;
-        public uint xWindowBorders;
-        public uint yWindowBorders;
-        public short atomWindowtype;
-        public short creatorVersion;
-    }
-
-    [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
-    public struct NOTIFYICONDATA
-    {
-        public uint cbSize;
-        public IntPtr hWnd;
-        public uint uID;
-        public uint uFlags;
-        public uint uCallbackMessage;
-        public IntPtr hIcon;
-        [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 128)]
-        public string szTip;
-        // campos posteriores ignorados para no recargar la estructura
+        public bool fIcon;
+        public uint xHotspot;
+        public uint yHotspot;
+        public IntPtr hbmMask;
+        public IntPtr hbmColor;
     }
 
     // Structure for menu-item information
@@ -76,5 +61,62 @@ internal static partial class NativeMethods
         public IntPtr dwTypeData;
         public uint cch;
         public IntPtr hbmpItem;
+    }
+
+    [StructLayout(LayoutKind.Sequential)]
+    public struct MSLLHOOKSTRUCT
+    {
+        public POINT pt;
+        public uint mouseData;
+        public uint flags;
+        public uint time;
+        public IntPtr dwExtraInfo;
+    }
+
+    [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
+    public struct NOTIFYICONDATA
+    {
+        public uint cbSize;
+        public IntPtr hWnd;
+        public uint uID;
+        public uint uFlags;
+        public uint uCallbackMessage;
+        public IntPtr hIcon;
+        [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 128)]
+        public string szTip;
+        // campos posteriores ignorados para no recargar la estructura
+    }
+
+    [StructLayout(LayoutKind.Sequential)]
+    public struct POINT
+    {
+        public int x;
+        public int y;
+    }
+
+    [StructLayout(LayoutKind.Sequential)]
+    public struct Rect
+    {
+        public int Left;
+        public int Top;
+        public int Right;
+        public int Bottom;
+
+        public readonly int Width => Right - Left;
+        public readonly int Height => Bottom - Top;
+    }
+
+    public struct WindowInfo
+    {
+        public int size;
+        public Rect window;
+        public Rect client;
+        public int style;
+        public int exStyle;
+        public int windowStatus;
+        public uint xWindowBorders;
+        public uint yWindowBorders;
+        public short atomWindowtype;
+        public short creatorVersion;
     }
 }
